chore(ui): fix clang-tidy debt surfaced by CI (Linux, WarningsAsErrors)
Local scripts/tidy.sh silently passed because run-clang-tidy on Windows
swallows diagnostics; CI invokes clang-tidy directly under
WarningsAsErrors: '*' and surfaced 348 errors across the Phase 52 UI +
tests. Filed separate todo to fix the script. This commit clears every
error in engine/ui/ and tests/unit/ surfaced by direct clang-tidy:

Bulk auto-fixes (clang-tidy --fix):
- readability-uppercase-literal-suffix:  u -> U, f -> F (many sites)
- readability-isolate-declaration:       multi-decl statements split

Manual fixes:
- nt_ui.c: extract dispatch_command() out of nt_ui_walk; the per-cmd
  switch was the bulk of the function's complexity. Walker drops from
  cognitive-complexity 66 to under the 25 threshold.
- nt_ui.c: NOLINTNEXTLINE on nt_ui_begin and nt_ui_set_atlas_white_region
  (engine convention for assert-heavy validation functions; matches
  precedent in nt_sprite_renderer.c).
- ui_atlas.c: NOLINTNEXTLINE on minimal_ui_atlas_create (test fixture
  builder with many pack-mount steps).
- ui_atlas.c: explicit size_t casts before pointer arithmetic to silence
  bugprone-implicit-widening-of-multiplication-result.
- test_nt_ui_lifecycle.c: NOLINTNEXTLINE on TEST_ASSERT_EQUAL_PTR (Unity
  macro casts through void * by contract).
- test_nt_sprite_renderer_emit_region.c: argument comment /* flip */
  -> /* flip_bits */ to match parameter name.
- ui_walker_fixture.h: bind mask type uint32_t instead of enum --
  callers compose bits via ~ALL & ~FLAG and the resulting values aren't
  enum members, which tripped clang-analyzer EnumCastOutOfRange.
- test_nt_ui_multictx.c: NOLINTNEXTLINE on the two interleave tests
  (multiple ctx loops by design).

46/46 ctest still green; clang-format clean; direct clang-tidy zero
errors across engine/ui/ + tests/unit/ (deps/stb noise filtered as
before by the project tidy.sh).

+++ tests/unit/test_helpers/ui_atlas.c
@@ -105,8 +105,8 @@ static void ui_atlas_build_inner_blob(uint8_t *out_blob) {
         .source_h = 1,
         .trim_offset_x = 0,
         .trim_offset_y = 0,
-        .origin_x = 0.0f,
-        .origin_y = 0.0f,
+        .origin_x = 0.0F,
+        .origin_y = 0.0F,
         .vertex_start = 0,
         .index_start = 0,
         .vertex_count = 4,
@@ -116,7 +116,7 @@ static void ui_atlas_build_inner_blob(uint8_t *out_blob) {
         .flags = NT_ATLAS_REGION_FLAG_QUAD_012023,
         ._reserved = {0, 0, 0},
     };
-    memcpy(out_blob + UI_ATLAS_HEADER_SIZE + UI_ATLAS_PAGE_IDS_SIZE, &region0, sizeof region0);
+    memcpy(out_blob + (size_t)UI_ATLAS_HEADER_SIZE + (size_t)UI_ATLAS_PAGE_IDS_SIZE, &region0, sizeof region0);
 
     /* ---- Region 1: 6-vertex polygon hull (12 indices = 4 fan triangles) ---- */
     NtAtlasRegion region1 = {
@@ -125,8 +125,8 @@ static void ui_atlas_build_inner_blob(uint8_t *out_blob) {
         .source_h = 16,
         .trim_offset_x = 0,
         .trim_offset_y = 0,
-        .origin_x = 0.5f,
-        .origin_y = 0.5f,
+        .origin_x = 0.5F,
+        .origin_y = 0.5F,
         .vertex_start = 4,
         .index_start = 6,
         .vertex_count = 6,
@@ -136,7 +136,7 @@ static void ui_atlas_build_inner_blob(uint8_t *out_blob) {
         .flags = 0,
         ._reserved = {0, 0, 0},
     };
-    memcpy(out_blob + UI_ATLAS_HEADER_SIZE + UI_ATLAS_PAGE_IDS_SIZE + 40, &region1, sizeof region1);
+    memcpy(out_blob + (size_t)UI_ATLAS_HEADER_SIZE + (size_t)UI_ATLAS_PAGE_IDS_SIZE + 40U, &region1, sizeof region1);
 
     /* ---- Vertices: 4 white + 6 polygon-hull ---- */
     NtAtlasVertex verts[UI_ATLAS_VERTEX_COUNT] = {
@@ -188,9 +188,9 @@ static uint8_t *ui_atlas_build_pack_blob(uint64_t atlas_rid, uint32_t *out_total
     ui_atlas_build_inner_blob(inner);
 
     const uint32_t raw_header = (uint32_t)(sizeof(NtPackHeader) + sizeof(NtAssetEntry));
-    const uint32_t header_size = (raw_header + (NT_PACK_DATA_ALIGN - 1u)) & ~(uint32_t)(NT_PACK_DATA_ALIGN - 1u);
+    const uint32_t header_size = (raw_header + (NT_PACK_DATA_ALIGN - 1U)) & ~(uint32_t)(NT_PACK_DATA_ALIGN - 1U);
     const uint32_t atlas_offset = header_size;
-    const uint32_t aligned_atlas = (UI_ATLAS_BLOB_SIZE + (NT_PACK_ASSET_ALIGN - 1u)) & ~(uint32_t)(NT_PACK_ASSET_ALIGN - 1u);
+    const uint32_t aligned_atlas = (UI_ATLAS_BLOB_SIZE + (NT_PACK_ASSET_ALIGN - 1U)) & ~(uint32_t)(NT_PACK_ASSET_ALIGN - 1U);
     const uint32_t total_size = atlas_offset + aligned_atlas;
 
     uint8_t *pack_blob = (uint8_t *)calloc(1, total_size);
@@ -221,6 +221,7 @@ static uint8_t *ui_atlas_build_pack_blob(uint64_t atlas_rid, uint32_t *out_total
     return pack_blob;
 }
 
+// NOLINTNEXTLINE(readability-function-cognitive-complexity)
 minimal_ui_atlas_t minimal_ui_atlas_create(void) {
     s_helper_counter++;
